Solved odd camera behavior and fixed events error.
Events errror:
dynarray_push was being called with the address of a shallow copy
of the immediate_observer array. Leaving the original pointer to
the arrays memory pointing to freed memory after reallocation

Camera behavior:
Camera's pitch was clamped between degrees -89 & 89. So when the
mouse first appears on the screen a large offset change is detected
changing the cameras pitch and yaw. Without restriction the camera
could turn upside down, showing an upside down image.